[4] add try-except and more posters

+++ intelli-analyze/_lab-4/_lab-4.py
@@ -1,4 +1,5 @@
 import json
+import os.path
 from bs4 import BeautifulSoup
 import requests
 import pandas as pd
@@ -16,11 +17,16 @@ data = pd.read_csv('../_lab-4/imdb_movies.csv', header=0,
         usecols=['id', 'genres', 'title', 'popularity', 'release_date', 'vote_average', 'vote_count', 'link'],
         dtype=object, parse_dates=True)
 
-for links, ids in zip(data['link'].tolist(), data['id'].tolist()):
-    r = requests.get(links)
-    #print('Connected!' if r.status_code==200 else 'Fail.')
-    soup = BeautifulSoup(r.text, 'html.parser')
-    im = requests.get(soup.find(class_='ipc-image').get('src'))
-    with open(path+ids+'.jpg', 'wb') as w:
-        w.write(im.content)
-        w.close()
+for links, ids in zip(data['link'][9696:].tolist(), data['id'][9696:].tolist()):
+    try:
+        if os.path.isfile(path+ids+'.jpg'):
+            continue
+        r = requests.get(links)
+        #print('Connected!' if r.status_code==200 else 'Fail.')
+        soup = BeautifulSoup(r.text, 'html.parser')
+        im = requests.get(soup.find(class_='ipc-image').get('src'))
+        with open(path+ids+'.jpg', 'wb') as w:
+            w.write(im.content)
+            w.close()
+    except Exception:
+        print(f'no link id: {ids}')
